Damage 처리 시 float 대신 int로 처리
Project : Client
[Modify]
M1. hjActor.h
M2. hjAttackScript
M3. hjBoss1BulletColliderScript.cpp
M4. hjBoss1LaserCollider.cpp
M5. hjBoss1SwordAttackCollider.cpp
M6. hjBossDeathParticleSystem.cpp
M7. hjStage1Boss

+++ Client/hjBoss1BulletColliderScript.cpp
@@ -30,7 +30,7 @@ namespace hj
 		if (nullptr == player)
 			return;
 
-		float test = 10.f;
+		int test = 10;
 		player->Damaged(test);
 
 		Animator* animator = mCircleCollider->GetOwner()->GetComponent<Animator>();
